[Fix] refactoring detail 페이지
