+++ src/redux/config/configStore.js
@@ -1,16 +1,16 @@
 import users from "../modules/users";
-import comments from "../modules/commentsSlice";
+import comments from "../modules/comments";
 import todos from "../modules/todos";
 import { configureStore } from "@reduxjs/toolkit";
 import { useSelector } from "react-redux";
 import { TypedUseSelectorHook, useDispatch } from "react-redux";
 
 const store = configureStore({
-    reducer: {
-        users,
-        comments,
-        todos
-    }
+  reducer: {
+    users,
+    comments,
+    todos,
+  },
 });
 
 export default store;
